+++ src/Entities/Schema.php
@@ -129,9 +129,6 @@ final class Schema implements IteratorAggregate, ArrayAccess
             'subscriptions/{subscriptionId}/interim-invoice' => function (array $content) {
                 return new Invoice($content);
             },
-            'subscriptions/{subscriptionId}/upcoming-invoices' => function (array $content) {
-                return new Collection(new Invoice(), $content);
-            },
             'subscriptions/{subscriptionId}/upcoming-invoice' => function (array $content) {
                 return new Invoice($content);
             },