+++ src/Client.php
@@ -105,7 +105,7 @@ final class Client
      */
     public const CURRENT_VERSION = 'v2.1';
 
-    public const SDK_VERSION = '2.22.0';
+    public const SDK_VERSION = '2.23.0';
 
     private static $services = [
         'authenticationOptions' => Services\AuthenticationOptionsService::class,